Refactoring function parseString
