--- app.js
@@ -1,6 +1,0 @@
-import timerFunc from './timerFunc.js';
-import { argv } from 'node:process';
-const timerTime = argv[2];
-
-timerFunc(timerTime);
-
