--- index.js
@@ -1,2 +1,0 @@
-import './app.js'
-//вводить значения таймера в формате строки 1h 20m 10s
