+++ package.json
@@ -2,7 +2,7 @@
   "name": "purple-school",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "timer/app.js",
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
